pip_install_wheel: Use --ignore-installed for pip
Replace the use of --force-reinstall with --ignore-installed when running pip
install. It can detect currently installed modules in the wrong environment and
try to remove them currently which is not what is desired. Ignoring them is the
correct thing to do.

(From OE-Core rev: 469380876044207d4537477d5dde93a4cf8e1e0c)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>